requirements.txt
google-adk[eval]
python-dotenv
aiohttp
pytest==9.0.1
pytest-asyncio==1.3.0
dateparser
python-dateutil
rapidfuzz>=3.0.0
